requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
groq
pydantic==2.9.0
python-dotenv==1.0.1
sqlalchemy==2.0.35
requests
gunicorn
